# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: WIAL Backend API
  version: 1.0.0
  description: |
    OpenAPI documentation for the WIAL backend.

    Public auth is login-only. User creation is restricted to authenticated
    administrators through the managed users endpoint.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid credentials
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users:
    get:
      tags: [Users]
      summary: List managed users
      description: Chapter leads are scoped to their assigned chapter. Super-admins can optionally filter by `chapter_id`.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Users]
      summary: Create a managed user
      description: |
        Creates a `chapter_lead`, `coach`, or `content_creator`.
        `super_admin` users are bootstrapped from env only and cannot be created through the API.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request, invalid role, or missing chapter assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: email already exists
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get a managed user
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Users]
      summary: Patch a managed user
      operationId: patchUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: User patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request, invalid role, or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: email already exists
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Delete a managed user
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/portal/overview:
    get:
      tags: [Portal]
      summary: Get leader portal overview
      description: Chapter leads are scoped to their assigned chapter. Super-admins must pass `chapter_id`.
      operationId: getPortalOverview
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
          description: Required for `super_admin`; ignored for `chapter_lead`.
      responses:
        '200':
          description: Portal overview for the scoped chapter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalOverviewResponse'
        '400':
          description: Missing required chapter_id for super-admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: chapter_id is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/portal/chapter:
    get:
      tags: [Portal]
      summary: Get the scoped portal chapter
      description: Chapter leads are scoped to their assigned chapter. Super-admins must pass `chapter_id`.
      operationId: getPortalChapter
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
          description: Required for `super_admin`; ignored for `chapter_lead`.
      responses:
        '200':
          description: Scoped chapter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: Missing required chapter_id for super-admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: chapter_id is required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/chapters:
    get:
      tags: [Chapters]
      summary: List chapters
      operationId: listChapters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: region
          schema:
            type: string
        - in: query
          name: language
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of chapters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Chapters]
      summary: Create a chapter
      description: Super-admin only. Each country can have only one chapter.
      operationId: createChapter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterCreateRequest'
      responses:
        '201':
          description: Chapter created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Chapter slug or country already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: country already exists
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/chapters/{id}:
    get:
      tags: [Chapters]
      summary: Get a chapter by ID
      operationId: getChapter
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Chapter found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Chapters]
      summary: Replace a chapter
      description: Each country can have only one chapter.
      operationId: updateChapter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterUpdateRequest'
      responses:
        '200':
          description: Chapter updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Chapter slug or country already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: country already exists
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Chapters]
      summary: Patch a chapter
      description: Super-admin global access; chapter leads are limited to their own chapter. Each country can have only one chapter.
      operationId: patchChapter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterPatchRequest'
      responses:
        '200':
          description: Chapter patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Chapter slug or country already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: country already exists
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Chapters]
      summary: Delete a chapter
      description: Super-admin global access; chapter leads are limited to their own chapter.
      operationId: deleteChapter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Chapter deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/chapters/{id}/content:
    patch:
      tags: [Chapters]
      summary: Patch chapter page content
      description: Content creators can edit only their own chapter. Super-admins can edit any chapter.
      operationId: patchChapterContent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterContentUpdateRequest'
      responses:
        '200':
          description: Chapter content updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/coaches:
    get:
      tags: [Coaches]
      summary: List coaches
      operationId: listCoaches
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: chapter_id
          schema:
            type: string
        - in: query
          name: certification_level
          schema:
            type: string
        - in: query
          name: language
          schema:
            type: string
        - in: query
          name: specialization
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of coaches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Coaches]
      summary: Create a coach profile
      description: Super-admin global access; chapter leads are limited to their own chapter.
      operationId: createCoach
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachCreateRequest'
      responses:
        '201':
          description: Coach created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coach'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/coaches/{id}:
    get:
      tags: [Coaches]
      summary: Get a coach by ID
      description: Public coach profile endpoint.
      operationId: getCoach
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Coach found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coach'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Coaches]
      summary: Patch a coach profile
      operationId: patchCoach
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachPatchRequest'
      responses:
        '200':
          description: Coach patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coach'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Coaches]
      summary: Delete a coach profile
      operationId: deleteCoach
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Coach deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events:
    get:
      tags: [Events]
      summary: List events
      operationId: listEvents
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: chapter_id
          schema:
            type: string
        - in: query
          name: event_type
          schema:
            type: string
      responses:
        '200':
          description: A paginated list of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Events]
      summary: Create an event
      description: Super-admin global access; chapter leads are limited to their own chapter.
      operationId: createEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/events/{id}:
    patch:
      tags: [Events]
      summary: Patch an event
      operationId: patchEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPatchRequest'
      responses:
        '200':
          description: Event patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Events]
      summary: Delete an event
      operationId: deleteEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Event deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/team-members:
    get:
      tags: [Team Members]
      summary: List chapter team members
      operationId: listTeamMembers
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
      responses:
        '200':
          description: Team member list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Team Members]
      summary: Create a chapter team member
      operationId: createTeamMember
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberCreateRequest'
      responses:
        '201':
          description: Team member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/team-members/{id}:
    get:
      tags: [Team Members]
      summary: Get a chapter team member
      operationId: getTeamMember
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Team member found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Team Members]
      summary: Patch a chapter team member
      operationId: patchTeamMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberPatchRequest'
      responses:
        '200':
          description: Team member patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '400':
          description: Invalid request or no fields to update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Team Members]
      summary: Delete a chapter team member
      operationId: deleteTeamMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Team member deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/resources:
    get:
      tags: [Resources]
      summary: List resources
      description: Omitting `chapter_id` returns global resources only.
      operationId: listResources
      parameters:
        - in: query
          name: chapter_id
          schema:
            type: string
      responses:
        '200':
          description: Resource list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Resources]
      summary: Create a resource
      description: Super-admins can create global or chapter resources. Chapter leads can create resources only for their chapter.
      operationId: createResource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceCreateRequest'
      responses:
        '201':
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/ai/coach-search:
    get:
      tags: [AI]
      summary: Search coaches using the simulated AI endpoint
      operationId: aiCoachSearch
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Coach search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AICoachSearchResponse'
        '400':
          description: Query parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks the required role or assignment scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
      required: [data]
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: '#/components/schemas/User'
      required: [token, user]
...